[level 3] Title: 등굣길, Time: 0.06 ms, Memory: 3.81 MB -BaekjoonHub

+++ 프로그래머스/lv3/42898. 등굣길/등굣길.cpp
@@ -1,7 +1,5 @@
 #include <string>
 #include <vector>
-#include <algorithm>
-#include <iostream>
 using namespace std;
 
 inline bool    is_puddles(const int &m, const int &n, const vector<vector<int>> &puddles) {
@@ -12,7 +10,7 @@ inline bool    is_puddles(const int &m, const int &n, const vector<vector<int>> 
 }
 
 int solution(int m, int n, vector<vector<int>> puddles) {
-    vector<vector<unsigned long long>> map(n, vector<unsigned long long>(m, 0));
+    vector<vector<int>> map(n, vector<int>(m, 0));
     for (int i = 0; i < puddles.size(); ++i)
         --puddles[i][0], --puddles[i][1];
     
@@ -21,18 +19,10 @@ int solution(int m, int n, vector<vector<int>> puddles) {
         for (int y = 0; y < n; ++y) {
             if (is_puddles(x, y, puddles)) continue ;
             int s = 0;
-            // if (y) map[y][x] += map[y - 1][x];
-            // if (x) map[y][x] += map[y][x - 1];
             if (y) s += map[y - 1][x];
             if (x) s += map[y][x - 1];
             map[y][x] += s % 1000000007;
         }
     }
-    // for_each(map.begin(), map.end(), [](auto v) {
-    //    for_each(v.begin(), v.end(), [](auto i) {
-    //        cout << i;
-    //    });
-    //     cout << endl;
-    // });
     return map[n - 1][m - 1];
 }
